Fix a bug in the app's repo lifecycle.
Previously, the lifecycle of a repo was:
  validating -> initializing -> hooks initializing -> ready

However, the "initialize" method (which corresponds to the
"initializing" state) was reused to update the repo in response to any
web hooks being called, and this caused it to continuously recreate the
web hook (as it thought it needed to transition from the "initializing"
to "hooks initializing" state).

This change fixes that bug by reordering the lifecycle states so that
"hooks initializing" comes *before* "initializing", and so the
"initialize" method does not need to do any follow up work other than
make sure the repo is in the "ready" state (which is an idempotent operation).