Fixed a bug in the detection of existing reviews.
Previously, this mirror was assuming that the review created from
a pull request was anchored at the first commit included in that
pull request. However, the base of a pull request can move forward
after it has been created, if the target ref is moved forward.

This means that the mirror should look at all existing reviews
with the same review ref, rather than just the existing review
requests at the (current) first commit in the pull request.